# api/web_page/openapi.yaml
swagger: "2.0"
info:
    title: Mai.Today Server
    description: Mai.Today Server
    version: 1.0.0
host: dev.api.mai.today
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /{channel}/createFolder:
        get:
            tags:
                - folder
            summary: '[WebSocket] Create a new folder'
            operationId: folder#receiveCreateFolder
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderReceiveCreateFolderOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /{channel}/deleteFolder:
        get:
            tags:
                - folder
            summary: '[WebSocket] Delete a folder by ID'
            operationId: folder#receiveDeleteFolder
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderReceiveDeleteFolderOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
    /{channel}/listFolders:
        get:
            tags:
                - folder
            summary: '[WebSocket] List all folders'
            operationId: folder#receiveListFolders
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderReceiveListFoldersOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /{channel}/updateFolder:
        get:
            tags:
                - folder
            summary: '[WebSocket] Update folder information by ID'
            operationId: folder#receiveUpdateFolder
            parameters:
                - name: channel
                  in: path
                  description: 使用者識別碼
                  required: true
                  type: string
                  maxLength: 24
                  minLength: 24
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "101":
                    description: 此 API 僅用於 WebSocket 協議
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderReceiveUpdateFolderOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
    /folders:
        get:
            tags:
                - folder
            summary: List all folders
            operationId: folder#ListFolders
            parameters:
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderListFoldersOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "400":
                    description: 參數錯誤
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/FolderListFoldersInvalidTokenResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
        post:
            tags:
                - folder
            summary: Create a new folder
            operationId: folder#CreateFolder
            parameters:
                - name: Authorization
                  in: header
                  required: false
                  type: string
                - name: CreateFolderRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FolderCreateFolderRequestBody'
                    required:
                        - folder
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderCreateFolderOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "400":
                    description: 參數錯誤
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/FolderCreateFolderInvalidTokenResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
    /folders/{folderId}:
        put:
            tags:
                - folder
            summary: Update folder information by ID
            operationId: folder#UpdateFolder
            parameters:
                - name: folderId
                  in: path
                  description: Folder identifier
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  required: false
                  type: string
                - name: UpdateFolderRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FolderUpdateFolderRequestBody'
                    required:
                        - folder
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderUpdateFolderOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "400":
                    description: 參數錯誤
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/FolderUpdateFolderInvalidTokenResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
        delete:
            tags:
                - folder
            summary: Delete a folder by ID
            operationId: folder#DeleteFolder
            parameters:
                - name: folderId
                  in: path
                  description: Folder identifier
                  required: true
                  type: string
                - name: Authorization
                  in: header
                  required: false
                  type: string
            responses:
                "200":
                    description: 操作成功
                    schema:
                        $ref: '#/definitions/FolderDeleteFolderOKResponseBody'
                        required:
                            - command
                            - timestamp
                            - data
                "401":
                    description: Unauthorized response.
                    schema:
                        $ref: '#/definitions/FolderDeleteFolderInvalidTokenResponseBody'
                "500":
                    description: 伺服器發生錯誤
            schemes:
                - https
            security:
                - JWT_header_Authorization: []
...
definitions:
...
    FolderCreateFolderRequestBody:
        title: FolderCreateFolderRequestBody
        type: object
        properties:
            folder:
                $ref: '#/definitions/FolderRequestBody'
        example:
            folder:
                baseId: base123
                createdAt: "2024-01-01T00:00:00Z"
                data:
                    color: red
                    name: My Folder
                folderId: "123456"
                parentId: parent123
                position: 1
                type: type123
                updatedAt: "2024-01-02T00:00:00Z"
        required:
            - folder
...
    FolderReceiveCreateFolderOKResponseBody:
        title: FolderReceiveCreateFolderOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/CreatefolderresultdataResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: createFolder
            data:
                baseId: base123
                createdAt: "2024-01-01T00:00:00Z"
                data:
                    color: red
                    name: My Folder
                folderId: "123456"
                parentId: parent123
                position: 1
                type: type123
                updatedAt: "2024-01-02T00:00:00Z"
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
    FolderReceiveDeleteFolderOKResponseBody:
        title: FolderReceiveDeleteFolderOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/CreatefolderresultdataResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: deleteFolder
            data:
                baseId: base123
                createdAt: "2024-01-01T00:00:00Z"
                data:
                    color: red
                    name: My Folder
                folderId: "123456"
                parentId: parent123
                position: 1
                type: type123
                updatedAt: "2024-01-02T00:00:00Z"
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
    FolderReceiveListFoldersOKResponseBody:
        title: FolderReceiveListFoldersOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                type: array
                items:
                    $ref: '#/definitions/CreatefolderresultdataResponseBody'
                description: 資料
                example:
                    - baseId: base123
                      createdAt: "2024-01-01T00:00:00Z"
                      data:
                        color: red
                        name: My Folder
                      folderId: "123456"
                      parentId: parent123
                      position: 1
                      type: type123
                      updatedAt: "2024-01-02T00:00:00Z"
                    - baseId: base123
                      createdAt: "2024-01-01T00:00:00Z"
                      data:
                        color: red
                        name: My Folder
                      folderId: "123456"
                      parentId: parent123
                      position: 1
                      type: type123
                      updatedAt: "2024-01-02T00:00:00Z"
                    - baseId: base123
                      createdAt: "2024-01-01T00:00:00Z"
                      data:
                        color: red
                        name: My Folder
                      folderId: "123456"
                      parentId: parent123
                      position: 1
                      type: type123
                      updatedAt: "2024-01-02T00:00:00Z"
                    - baseId: base123
                      createdAt: "2024-01-01T00:00:00Z"
                      data:
                        color: red
                        name: My Folder
                      folderId: "123456"
                      parentId: parent123
                      position: 1
                      type: type123
                      updatedAt: "2024-01-02T00:00:00Z"
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: listFolders
            data:
                - baseId: base123
                  createdAt: "2024-01-01T00:00:00Z"
                  data:
                    color: red
                    name: My Folder
                  folderId: "123456"
                  parentId: parent123
                  position: 1
                  type: type123
                  updatedAt: "2024-01-02T00:00:00Z"
                - baseId: base123
                  createdAt: "2024-01-01T00:00:00Z"
                  data:
                    color: red
                    name: My Folder
                  folderId: "123456"
                  parentId: parent123
                  position: 1
                  type: type123
                  updatedAt: "2024-01-02T00:00:00Z"
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
    FolderReceiveUpdateFolderOKResponseBody:
        title: FolderReceiveUpdateFolderOKResponseBody
        type: object
        properties:
            command:
                $ref: '#/definitions/CommandResponseBody'
            data:
                $ref: '#/definitions/CreatefolderresultdataResponseBody'
            timestamp:
                type: integer
                description: 時間戳記
                example: 1719188007000000000
                format: int64
                minimum: 1.719188007e+18
        example:
            command:
                type: updateFolder
            data:
                baseId: base123
                createdAt: "2024-01-01T00:00:00Z"
                data:
                    color: red
                    name: My Folder
                folderId: "123456"
                parentId: parent123
                position: 1
                type: type123
                updatedAt: "2024-01-02T00:00:00Z"
            timestamp: 1719188007000000000
        required:
            - command
            - timestamp
            - data
...
securityDefinitions:
    JWT_header_Authorization:
        type: apiKey
        name: Authorization
        in: header